requirements.txt
python-telegram-bot
rembg
pillow
pytz